home work delete watch roles

+++ TechNotes.Application/Notes/DeleteNote/DeleteNoteCommandHandler.cs
@@ -1,15 +1,25 @@
+using TechNotes.Application.Users;
+
 namespace TechNotes.Application.Notes.DeleteNote;
 
 public class DeleteNoteCommandHandler : ICommandHandler<DeleteNoteCommand>
 {
     private readonly INoteRepository _noteRepository;
+    private readonly IUserService _userService;
 
-    public DeleteNoteCommandHandler(INoteRepository noteRepository)
+    public DeleteNoteCommandHandler(INoteRepository noteRepository, IUserService userService)
     {
         _noteRepository = noteRepository;
+        _userService = userService;
     }
     public async Task<Result> Handle(DeleteNoteCommand request, CancellationToken cancellationToken)
     {
+        var currentUserCanDelete = await _userService.CurrentUserCanEditNoteAsync(request.Id);
+        if (!currentUserCanDelete)
+        {
+            return Result.Fail("You are not the owner of this note.");
+        }
+
         var deleted = await _noteRepository.DeleteNoteAsync(request.Id);
 
         if (deleted)
